+++ packages/util-crypto/src/noWasm.ts
@@ -2,10 +2,12 @@
 // This software may be modified and distributed under the terms
 // of the Apache-2.0 license. See the LICENSE file for details.
 
-export function cryptoIsReady (): boolean {
-  return true;
+export function isReady (): boolean {
+  // always false, when true it will try and use non-existent functions
+  return false;
 }
 
-export function cryptoWaitReady (): Promise<boolean> {
+export function waitReady (): Promise<boolean> {
+  // always immediate true, our process is done
   return Promise.resolve(true);
 }